// api/models/PostUpdateRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 更新帖子请求
 */
export type PostUpdateRequest = {
    /**
     * 帖子ID
     */
    communityPostId: number;
    /**
     * 帖子标题
     */
    title?: string;
    /**
     * 帖子内容
     */
    content?: string;
};