// api/models/CommunityMemberRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 社区成员请求
 */
export type CommunityMemberRequest = {
    /**
     * 社区ID
     */
    communityId: number;
    /**
     * 用户ID
     */
    userId: number;
};